requirements.txt
google-cloud-storage==2.16.0
matplotlib==3.9.0
numpy==1.26.4
streamlit==1.35.0
